+++ test/rules/test_plugin_manager.py
@@ -1315,6 +1315,7 @@ def test_markdown_with_plugins_list_only(scanner_default: MarkdownScanner) -> No
   md051   link-fragments                  True       True       0.5.0    No
   md053   link-image-reference-definitio  True       True       0.5.0    No
           ns
+  md054   link-image-style                True       True       0.5.0    No
   md059   descriptive-link-text           True       True       0.5.0    No
   pml100  disallowed-html                 False      False      0.6.0    No
   pml101  list-anchored-indent            False      False      0.6.0    No
@@ -1403,6 +1404,7 @@ def test_markdown_with_plugins_list_only_all(scanner_default: MarkdownScanner) -
   md051   link-fragments                  True       True       0.5.0    No
   md053   link-image-reference-definitio  True       True       0.5.0    No
           ns
+  md054   link-image-style                True       True       0.5.0    No
   md059   descriptive-link-text           True       True       0.5.0    No
   md999   debug-only                      False      False      0.0.0    No
   pml100  disallowed-html                 False      False      0.6.0    No
@@ -1493,6 +1495,7 @@ def test_markdown_with_plugins_list_after_command_line_disable_all_rules(
   md051   link-fragments                  True       False      0.5.0    No
   md053   link-image-reference-definitio  True       False      0.5.0    No
           ns
+  md054   link-image-style                True       False      0.5.0    No
   md059   descriptive-link-text           True       False      0.5.0    No
   pml100  disallowed-html                 False      False      0.6.0    No
   pml101  list-anchored-indent            False      False      0.6.0    No
@@ -1587,6 +1590,7 @@ def test_markdown_with_plugins_list_after_configuration_disable_all_rules(
   md051   link-fragments                  True       False      0.5.0    No
   md053   link-image-reference-definitio  True       False      0.5.0    No
           ns
+  md054   link-image-style                True       False      0.5.0    No
   md059   descriptive-link-text           True       False      0.5.0    No
   pml100  disallowed-html                 False      False      0.6.0    No
   pml101  list-anchored-indent            False      False      0.6.0    No
@@ -1683,6 +1687,7 @@ def test_markdown_with_plugins_list_after_command_line_disable_all_rules_and_ena
   md051   link-fragments                  True       False      0.5.0    No
   md053   link-image-reference-definitio  True       False      0.5.0    No
           ns
+  md054   link-image-style                True       False      0.5.0    No
   md059   descriptive-link-text           True       False      0.5.0    No
   pml100  disallowed-html                 False      False      0.6.0    No
   pml101  list-anchored-indent            False      False      0.6.0    No